refactor : customer 도메인 추가로 인한 패키지 구조 변경
