+++ src/main/java/com/prgms/VoucherApp/controller/VoucherApp.java
@@ -1,10 +1,10 @@
 package com.prgms.VoucherApp.controller;
 
-import com.prgms.VoucherApp.domain.Voucher;
-import com.prgms.VoucherApp.domain.VoucherType;
-import com.prgms.VoucherApp.dto.VoucherDto;
-import com.prgms.VoucherApp.model.VoucherCreator;
-import com.prgms.VoucherApp.model.VoucherReader;
+import com.prgms.VoucherApp.domain.voucher.Voucher;
+import com.prgms.VoucherApp.domain.voucher.VoucherType;
+import com.prgms.VoucherApp.domain.voucher.dto.VoucherDto;
+import com.prgms.VoucherApp.domain.voucher.model.VoucherCreator;
+import com.prgms.VoucherApp.domain.voucher.model.VoucherReader;
 import com.prgms.VoucherApp.view.Command;
 import com.prgms.VoucherApp.view.Input;
 import com.prgms.VoucherApp.view.Output;
