+++ src/main/java/com/prgms/VoucherApp/view/Output.java
@@ -1,8 +1,8 @@
 package com.prgms.VoucherApp.view;
 
-import com.prgms.VoucherApp.domain.Voucher;
-import com.prgms.VoucherApp.domain.VoucherType;
-import com.prgms.VoucherApp.dto.VoucherDto;
+import com.prgms.VoucherApp.domain.voucher.Voucher;
+import com.prgms.VoucherApp.domain.voucher.VoucherType;
+import com.prgms.VoucherApp.domain.voucher.dto.VoucherDto;
 
 import java.util.List;
