+++ src/main/java/com/prgms/VoucherApp/view/console/ConsoleOutputView.java
@@ -1,8 +1,8 @@
 package com.prgms.VoucherApp.view.console;
 
-import com.prgms.VoucherApp.domain.Voucher;
-import com.prgms.VoucherApp.domain.VoucherType;
-import com.prgms.VoucherApp.dto.VoucherDto;
+import com.prgms.VoucherApp.domain.voucher.Voucher;
+import com.prgms.VoucherApp.domain.voucher.VoucherType;
+import com.prgms.VoucherApp.domain.voucher.dto.VoucherDto;
 import com.prgms.VoucherApp.view.Output;
 import org.beryx.textio.TextIO;
 import org.beryx.textio.TextIoFactory;
