+++ src/test/java/com/prgms/VoucherApp/voucher/domain/VoucherTest.java
@@ -1,5 +1,8 @@
-package com.prgms.VoucherApp.domain;
+package com.prgms.VoucherApp.voucher.domain;
 
+import com.prgms.VoucherApp.domain.voucher.FixedAmountVoucher;
+import com.prgms.VoucherApp.domain.voucher.PercentDiscountVoucher;
+import com.prgms.VoucherApp.domain.voucher.Voucher;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
