+++ src/test/java/com/prgms/VoucherApp/voucher/storage/VoucherFileStorageTest.java
@@ -1,25 +1,27 @@
-package com.prgms.VoucherApp.storage;
+package com.prgms.VoucherApp.voucher.storage;
 
-import com.prgms.VoucherApp.domain.FixedAmountVoucher;
-import com.prgms.VoucherApp.domain.PercentDiscountVoucher;
-import com.prgms.VoucherApp.domain.Voucher;
-import org.junit.jupiter.api.BeforeEach;
+import com.prgms.VoucherApp.domain.voucher.FixedAmountVoucher;
+import com.prgms.VoucherApp.domain.voucher.PercentDiscountVoucher;
+import com.prgms.VoucherApp.domain.voucher.Voucher;
+import com.prgms.VoucherApp.domain.voucher.storage.VoucherFileStorage;
+import com.prgms.VoucherApp.domain.voucher.storage.VoucherStorage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 import java.math.BigDecimal;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class VoucherMemoryStorageTest {
+@SpringBootTest(classes = {VoucherFileStorage.class})
+@ActiveProfiles(profiles = "dev")
+public class VoucherFileStorageTest {
 
+    @Autowired
     VoucherStorage storage;
-
-    @BeforeEach
-    void setUp() {
-        storage = new VoucherMemoryStorage();
-    }
 
     @Test
     @DisplayName("고정 비용 할인권 생성 테스트")
@@ -36,7 +38,7 @@ class VoucherMemoryStorageTest {
     @DisplayName("퍼센트 비율 할인권 생성 테스트")
     void savePercentVoucher() {
 
-        Voucher percentVoucher = new PercentDiscountVoucher(UUID.randomUUID(), BigDecimal.valueOf(1000));
+        Voucher percentVoucher = new PercentDiscountVoucher(UUID.randomUUID(), BigDecimal.valueOf(50));
 
         storage.save(percentVoucher);
